Create Debouncing pushbuttons in software
Debouncing 4 buttons in software. 
With the NUCLEO board L412KB four keys are debounced in the software. With each keystroke an external interrupt is triggered. A valid key press is checked in the SysTick handler (void SysTick_Handler(void)) every ms after this event. If a stable state for at least 20ms is detected for the pressed key, the keystroke is recognized as valid in the software and a LED is toggled.